Refactored user data and di layer into the "core" package, as both schedule and profile will be using it
