+++ app/src/main/java/com/jikokujo/core/data/remote/ApiResult.kt
@@ -1,4 +1,4 @@
-package com.jikokujo.core.data
+package com.jikokujo.core.data.remote
 
 sealed interface ApiResult<out T>{
     data class Success<T>(val data: T) : ApiResult<T>
