+++ app/src/main/java/com/jikokujo/core/data/remote/ApiUserResponseStructures.kt
@@ -1,15 +1,14 @@
-package com.jikokujo.profile.data.remote
+package com.jikokujo.core.data.remote
 
 import com.google.gson.annotations.SerializedName
-import com.jikokujo.core.data.Payload
-import com.jikokujo.profile.data.model.User
+import com.jikokujo.core.data.model.User
 
 data class UserLoginObj(
     @SerializedName("token")
     val userAccessToken: String
-): Payload
+) : Payload
 
 data class GetUserObj(
     @SerializedName("user")
     val user: User
-): Payload
+) : Payload
