+++ app/src/main/java/com/jikokujo/profile/presentation/ProfileViewModel.kt
@@ -1,12 +1,11 @@
 package com.jikokujo.profile.presentation
 
-import androidx.compose.runtime.mutableStateListOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.navigation3.runtime.NavKey
-import com.jikokujo.core.data.ApiResult
-import com.jikokujo.profile.data.model.User
-import com.jikokujo.profile.data.repository.UserRepository
+import com.jikokujo.core.data.remote.ApiResult
+import com.jikokujo.core.data.model.User
+import com.jikokujo.core.data.repository.UserRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
