+++ app/src/main/java/com/jikokujo/profile/presentation/auth/RegisterViewModel.kt
@@ -1,9 +1,9 @@
 package com.jikokujo.profile.presentation.auth
 
 import androidx.lifecycle.ViewModel
-import com.jikokujo.core.data.ApiResult
+import com.jikokujo.core.data.remote.ApiResult
 import com.jikokujo.core.utils.validateEmail
-import com.jikokujo.profile.data.repository.UserRepository
+import com.jikokujo.core.data.repository.UserRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
