+++ app/src/main/java/com/jikokujo/schedule/data/repository/QueryablesRepository.kt
@@ -1,6 +1,6 @@
 package com.jikokujo.schedule.data.repository
 
-import com.jikokujo.core.data.ApiResult
+import com.jikokujo.core.data.remote.ApiResult
 import com.jikokujo.schedule.data.model.Queryable
 
 interface QueryablesRepository {
