+++ app/src/main/java/com/jikokujo/schedule/data/repository/QueryablesRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.jikokujo.schedule.data.repository
 
 import android.util.Log
-import com.jikokujo.core.data.ApiResult
+import com.jikokujo.core.data.remote.ApiResult
 import com.jikokujo.schedule.data.model.Queryable
 import com.jikokujo.schedule.data.remote.QueryablesApi
