+++ app/src/test/java/com/jikokujo/schedule/data/MockQueryablesRepository.kt
@@ -1,7 +1,7 @@
 package com.jikokujo.schedule.data
 
 import com.jikokujo.schedule.data.model.Queryable
-import com.jikokujo.core.data.ApiResult
+import com.jikokujo.core.data.remote.ApiResult
 import com.jikokujo.schedule.data.repository.QueryablesRepository
 
 class MockQueryablesRepositoryImpl(queryablesIn: List<Queryable>): QueryablesRepository {
