+++ app/src/test/java/com/jikokujo/schedule/data/MockTripsRepository.kt
@@ -4,7 +4,7 @@ import com.jikokujo.schedule.data.model.Location
 import com.jikokujo.schedule.data.model.Queryable
 import com.jikokujo.schedule.data.model.StopWithLocationAndStopTime
 import com.jikokujo.schedule.data.model.Trip
-import com.jikokujo.core.data.ApiResult
+import com.jikokujo.core.data.remote.ApiResult
 import com.jikokujo.schedule.data.model.RoutePathPoint
 import com.jikokujo.schedule.data.repository.TripsRepository
 import java.time.LocalDateTime
